create_ticket/requirements.txt
google-cloud-datastore==1.15.0
google-cloud-secret-manager==1.0.0
jira==2.0.0
retry==0.9.2
